refactor(navigation): create new folder and restructure layout
