+++ src/app/(pwa)/layout.tsx
@@ -1,10 +1,10 @@
-import Navbar from '@/ui/components/navbar';
+import BottomNav from '@/ui/components/navigation/bottom-nav';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <>
-      <Navbar />
       <main className='container pb-24'>{children}</main>
+      <BottomNav />
     </>
   );
 }
